Pulled shared functionality into CombatEntity
